feat: add custom_data_audit.py module, trim eda, fix imports
