+++ notebooks/custom_templates/README.md
@@ -1,5 +1,23 @@
-## SST Custom Templates 
-- We start our training notebook at `02-train-h2o-model-TEMPLATE` to assume that all preprocessing is 01, though these notebooks will likely be renumbered depending on the number of preprocessing notebooks before.
-- Define the _run_type_ parameter in a DB workflow setting with either "train" or "predict". These can be used in a workflow once inference is performed or training.
-- This is the skeleton for a pipeline, but can be built upon depending on the amount of customization in the school's pipeline, processing, outcome definition etc.
-- 1 config is mapped to 1 model. If a school has multiple models, we need a config for each. This is so that model cards can be created accordingly and configs can have a flat & simple structure (same implemention for PDP).
+## SST custom templates
+
+- Define the _run_type_ parameter in a Databricks workflow as `"train"` or `"predict"` for notebooks that branch on that flag.
+- One `config.toml` maps to one model. Multiple models need multiple configs (same pattern as PDP).
+
+### Suggested notebook order
+
+| Step | Template | Role |
+|------|----------|------|
+| 00 | `00-data-audit-TEMPLATE.py` | Table-only institutional audit (keys, credits, merges). |
+| 01 | `01-data-assessment-TEMPLATE.py` | Charts / deeper EDA on bronze. |
+| 02 | `02-preprocess-data-TEMPLATE.py` | Silver build + schema contract via `custom_cleaning`. |
+| 03–06 | training, registration, predictions, validation | Downstream modeling. |
+
+Numbers on files (e.g. `06` before `03`) reflect historical naming; follow the pipeline role in the table above.
+
+### `edvise.data_audit` imports
+
+- **`eda`** — exploratory only: logging helpers, `analyze_merge`, `value_counts_*`, summaries. Does not own column inference or cleaning transforms.
+- **`custom_data_audit`** — structured audit: `find_dupes`, `validate_credit_consistency`, `infer_*` column pickers, pass/fail vs grade checks, etc.
+- **`custom_cleaning`** — transforms: `order_terms`, `normalize_student_id_column`, `clean_bronze_datasets`, schema contract helpers, etc.
+
+`eda` still re-exports many `custom_data_audit` / `custom_cleaning` names for older notebooks; new code should import from the submodule directly (as in templates 00 and 01).
